+++ src/Stack/ConfigCmd.hs
@@ -67,7 +67,7 @@ cfgCmdSet cmd = do
                      case mstackYaml of
                          PCProject stackYaml -> return stackYaml
                          PCGlobalProject -> liftM (</> stackDotYaml) (getImplicitGlobalProjectDir conf)
-                         PCNoConfig _extraDeps -> throwString "config command used when no local configuration available"
+                         PCNoProject _extraDeps -> throwString "config command used when no project configuration available" -- maybe modify the ~/.stack/config.yaml file instead?
                  CommandScopeGlobal -> return (configUserConfigPath conf)
     -- We don't need to worry about checking for a valid yaml here
     (config :: Yaml.Object) <-