+++ src/Stack/Script.hs
@@ -68,7 +68,7 @@ scriptCmd opts = do
         fromString (toFilePath fp)
       SYLGlobalProject -> logError "Ignoring SYLGlobalProject for script command"
       SYLDefault -> return ()
-      SYLNoConfig _ -> assert False (return ())
+      SYLNoProject _ -> assert False (return ())
 
     file <- resolveFile' $ soFile opts
     let scriptDir = parent file
@@ -76,7 +76,7 @@ scriptCmd opts = do
             { globalConfigMonoid = (globalConfigMonoid go)
                 { configMonoidInstallGHC = FirstTrue $ Just True
                 }
-            , globalStackYaml = SYLNoConfig $ soScriptExtraDeps opts
+            , globalStackYaml = SYLNoProject $ soScriptExtraDeps opts
             }
 
     -- Optimization: if we're compiling, and the executable is newer